+++ CHANGELOG.md
@@ -6,6 +6,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 ### Unreleased
 * Notifications
    * Design updates
+   * Delete notification
 
 ## 0.8.0